subscription_4

+++ backend/foodgram/api/views.py
@@ -110,49 +110,48 @@ class RecipesViewSet(viewsets.ModelViewSet):
         user = self.request.user
         recipe = get_object_or_404(Recipes, id=pk)
         if request.method == 'POST':
-            if user.is_authenticated:
-                if Favorite.objects.filter(
-                    recipes=recipe, user=user
-                ).exists():
-                    raise ValueError('Такоу рецепт уже существует')
-                Favorite.objects.create(recipes=recipe, user=user)
-                serializer = ShortViewRecipesSerializer(
-                    recipe, context={'request': request})
-                return Response(serializer.data,
-                                status=status.HTTP_201_CREATED)
-            return Response(status=status.HTTP_401_UNAUTHORIZED)
-        favorited = get_object_or_404(Favorite, recipes=recipe,
-                                      user=user)
-        favorited.delete()
-        # Favorite.objects.filter(recipes=recipe, user=user).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            if Favorite.objects.filter(
+                recipes=recipe, user=user
+            ).exists():
+                raise ValueError('Такоу рецепт уже существует')
+            Favorite.objects.create(recipes=recipe, user=user)
+            serializer = ShortViewRecipesSerializer(
+                recipe, context={'request': request})
+            return Response(serializer.data,
+                            status=status.HTTP_201_CREATED)
+        if request.method == 'DELETE':
+            favorited = get_object_or_404(Favorite, recipes=recipe,
+                                          user=user)
+            favorited.delete()
+            # Favorite.objects.filter(recipes=recipe, user=user).delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        # return Response(status=status.HTTP_400UNAUTHORIZED)
 
     @action(detail=True, serializer_class=ShortViewRecipesSerializer,
             methods=['post', 'delete'], permission_classes=[AuthorOnly])
     def shopping_cart(self, request, pk):
         user = self.request.user
         recipe = get_object_or_404(Recipes, id=pk)
-        if user.is_authenticated:
-            if request.method == "POST":
-                if Shopping_Cart.objects.filter(user=user,
-                                                recipes=recipe).exists():
-                    raise ValueError(
-                        'Такой рецепт уже добавлен в Ваш список покупок.'
-                    )
-                Shopping_Cart.objects.create(user=user, recipes=recipe)
-                serializer = ShortViewRecipesSerializer(
-                    recipe, context={'request': request}
+        if request.method == "POST":
+            if Shopping_Cart.objects.filter(user=user,
+                                            recipes=recipe).exists():
+                raise ValueError(
+                    'Такой рецепт уже добавлен в Ваш список покупок.'
                 )
-                return Response(serializer.data,
-                                status=status.HTTP_201_CREATED)
-            if not Shopping_Cart.objects.filter(user=user,
-                                                recipes=recipe).exists():
-                raise ValueError('Такого рецепта нет.')
-            # cart = get_object_or_404(Shopping_Cart, recipes=recipe,user=user)
+            Shopping_Cart.objects.create(user=user, recipes=recipe)
+            serializer = ShortViewRecipesSerializer(
+                recipe, context={'request': request}
+            )
+            return Response(serializer.data,
+                            status=status.HTTP_201_CREATED)
+        if not Shopping_Cart.objects.filter(user=user,
+                                            recipes=recipe).exists():
+            raise ValueError('Такого рецепта нет.')
+            # cart = get_object_or_404(Shopping_Cart,
+            # recipes=recipe,user=user)
             # cart.delete()
-            Shopping_Cart.objects.filter(recipes=recipe, user=user).delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status=status.HTTP_401_UNAUTHORIZED)
+        Shopping_Cart.objects.filter(recipes=recipe, user=user).delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, serializer_class=ShortViewRecipesSerializer,
             permission_classes=[IsAuthenticated, ], methods=['GET'])
